Changed the variable name 'best_j' to 'best_j_score'
In line 95 the earlier code was checking the 'if' condition between `score[best_i]` and `score[best_j]` and it was giving an error as 'best_j' was not defined. Since line 94 was using 'best_j_score' so I changed the variable name from 'best_j' to 'best_j_score'.